[Normal] Update Some Code

+++ tool.py
@@ -1620,6 +1620,9 @@ class dbkxyt:
             os.rename(dir_ + "firmware-update", dir_ + "images")
         if not os.path.exists(dir_ + "images"):
             os.makedirs(dir_ + 'images')
+        if os.path.exists(os.path.join(rwork(), 'payload.bin')):
+            print("Find payload.bin ,Stop!")
+            return
         if os.path.exists(dir_ + 'META-INF'):
             rmdir(dir_ + 'META-INF')
         zipfile.ZipFile(elocal + os.sep + "bin" + os.sep + "extra_flash.zip").extractall(dir_)
